Refactor: Move add-property page to client component
