+++ app/package.json
@@ -4,6 +4,7 @@
   "version": "0.1.0",
   "scripts": {
     "dev": "next dev -p 3000",
+    "prebuild": "node ../scripts/ensure-next-public-env.mjs",
     "build": "next build",
     "start": "next start -p 3000",
     "lint": "next lint"
